Only remove the cron.d file if it exists

+++ s6/debian-root/etc/cont-init.d/20-start.sh
@@ -18,7 +18,7 @@ if [ -z "$SKIPGRAVITYONBOOT" ]; then
     echo '@reboot root PATH="$PATH:/usr/sbin:/usr/local/bin/" pihole updateGravity >/var/log/pihole_updateGravity.log || cat /var/log/pihole_updateGravity.log' > /etc/cron.d/gravity-on-boot
 else
     echo "  Skipping Gravity Database Update."
-    rm /etc/cron.d/gravity-on-boot &>/dev/null
+    [ ! -e /etc/cron.d/gravity-on-boot ] || rm /etc/cron.d/gravity-on-boot &>/dev/null
 fi
 
 # Kill dnsmasq because s6 won't like it if it's running when s6 services start
